@@ -99,27 +99,27 @@ python = [
 ]
 
 [tool.hatch.envs.test]
-python = "3.7"
+python = "3.8"
 
 [tool.pyright]
-include = ["boario"]
+include = ["boario","scripts"]
 exclude = ["**/node_modules",
     "**/__pycache__",
-    "src/"
+    "src/",
+    "api-examples"
 ]
 ignore = ["src/oldstuff"]
 defineConstant = { DEBUG = true }
 stubPath = "boario/stubs"
+pythonPath = "/home/sjuhel/mambaforge-pypy3/envs/boario-dev/bin/python"
+venvPath = "/home/sjuhel/mambaforge-pypy3/envs/boario-dev/"
 
 reportMissingImports = true
 reportMissingTypeStubs = false
 
-pythonVersion = "3.6"
+pythonVersion = "3.8"
 pythonPlatform = "Linux"
 
 executionEnvironments = [
-  { root = "src/web", pythonVersion = "3.5", pythonPlatform = "Windows", extraPaths = [ "src/service_libs" ] },
-  { root = "src/sdk", pythonVersion = "3.0", extraPaths = [ "src/backend" ] },
-  { root = "src/tests", extraPaths = ["src/tests/e2e", "src/sdk" ]},
   { root = "src" }
 ]
