@@ -16,19 +16,7 @@
 
 from boario.event import Event
 from collections.abc import Iterable
-
-from json import JSONEncoder
-class EventEncoder(JSONEncoder):
-    def default(self, event:Event):
-        dic = dict(event.__dict__)
-        del dic["final_demand_rebuild"]
-        del dic["final_demand_rebuild_share"]
-        del dic["industry_rebuild"]
-        del dic["industry_rebuild_share"]
-        del dic["production_share_allocated"]
-        del dic["rebuildable"]
-        return dic
-
+import pymrio
 
 def flatten(l):
     for el in l:
@@ -36,3 +24,39 @@ def flatten(l):
             yield from flatten(el)
         else:
             yield el
+
+def lexico_reindex(mrio: pymrio.IOSystem) -> pymrio.IOSystem:
+    """Reindex IOSystem lexicographicaly
+
+    Sort indexes and columns of the dataframe of a ``pymrio`` `IOSystem <https://pymrio.readthedocs.io/en/latest/intro.html>` by
+    lexical order.
+
+    Parameters
+    ----------
+    mrio : pymrio.IOSystem
+        The IOSystem to sort
+
+    Returns
+    -------
+    pymrio.IOSystem
+        The sorted IOSystem
+
+
+    """
+    if mrio.Z is None:
+        raise ValueError("Given mrio has no Z attribute set")
+    mrio.Z = mrio.Z.reindex(sorted(mrio.Z.index), axis=0)
+    mrio.Z = mrio.Z.reindex(sorted(mrio.Z.columns), axis=1)
+    if mrio.Y is None:
+        raise ValueError("Given mrio has no Z attribute set")
+    mrio.Y = mrio.Y.reindex(sorted(mrio.Y.index), axis=0)
+    mrio.Y = mrio.Y.reindex(sorted(mrio.Y.columns), axis=1)
+    if mrio.x is None:
+        raise ValueError("Given mrio has no Z attribute set")
+    mrio.x = mrio.x.reindex(sorted(mrio.x.index), axis=0)
+    if mrio.A is None:
+        raise ValueError("Given mrio has no Z attribute set")
+    mrio.A = mrio.A.reindex(sorted(mrio.A.index), axis=0)
+    mrio.A = mrio.A.reindex(sorted(mrio.A.columns), axis=1)
+
+    return mrio
